used atof insted of converting price into cents. fixed rounding problems

+++ CPP_09/ex00/BitcoinExchange.hpp
@@ -14,8 +14,8 @@ class BitcoinExchange
         std::ifstream inputFile;
         std::ifstream fileDB;
         std::string tempDate;
-        int tempAmount;
-        std::map<std::string, int> DB;
+        double tempAmount;
+        std::map<std::string, double> DB;
     
     public:
         BitcoinExchange();
